update search apps/accounts categories
