+++ src/components/Search/components/SearchSuggestions/SearchResults/SearchAccounts/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchAccounts';
